// packages/print-toolkit/src/orca-schema.generated.ts
export const ORCA_SCHEMA: OrcaSchema = {
 "orcaVersion": "2.3.2",
 "options": {
// ...
  "filament_retraction_length": {
   "type": "floats",
   "label": "Length",
   "fullLabel": "Retraction Length",
   "tooltip": "Some amount of material in extruder is pulled back to avoid ooze during long travel. Set zero to disable retraction.",
   "sidetext": "mm",
   "nullable": true,
   "mode": "simple",
   "default": [
    "0.8"
   ]
  },
// ...
  "filament_type": {
   "type": "strings",
   "label": "Type",
   "tooltip": "The material type of filament.",
   "guiType": "f_enum_open",
   "mode": "simple",
   "default": [
    "PLA"
   ]
  },
// ...
  "gcode_flavor": {
   "type": "enum",
   "label": "G-code flavor",
   "tooltip": "What kind of G-code the printer is compatible with.",
   "enumValues": [
    "marlin",
    "klipper",
    "reprapfirmware",
    "marlin2"
   ],
   "enumLabels": [
    "Marlin(legacy)",
    "Klipper",
    "RepRapFirmware",
    "Marlin 2"
   ],
   "mode": "advanced",
   "default": "marlin"
  },
// ...
  "machine_max_speed_x": {
   "type": "floats",
   "fullLabel": "Maximum speed X",
   "label": "Maximum speed X",
   "category": "Machine limits",
   "tooltip": "Maximum speed of X axis",
   "sidetext": "mm/s",
   "min": 0,
   "mode": "simple",
   "default": [
    "500",
    "200"
   ]
  },
// ...
  "nozzle_diameter": {
   "type": "floats",
   "label": "Nozzle diameter",
   "tooltip": "The diameter of nozzle.",
   "sidetext": "mm",
   "mode": "advanced",
   "max": 100,
   "default": [
    "0.4"
   ]
  },
// ...
  "nozzle_temperature": {
   "type": "ints",
   "label": "Other layers",
   "tooltip": "Nozzle temperature for layers after the initial one.",
   "sidetext": "℃",
   "fullLabel": "Nozzle temperature",
   "min": 0,
   "default": [
    "200"
   ]
  },
// ...
  "printable_area": {
   "type": "points",
   "label": "Printable area",
   "mode": "advanced",
   "default": [
    "0x0",
    "200x0",
    "200x200",
    "0x200"
   ]
  },
// ...
  "retraction_length": {
   "type": "floats",
   "label": "Length",
   "fullLabel": "Retraction Length",
   "tooltip": "Some amount of material in extruder is pulled back to avoid ooze during long travel. Set zero to disable retraction.",
   "sidetext": "mm",
   "mode": "simple",
   "default": [
    "0.8"
   ]
  },
// ...
 },
// ...
};